player function completes
