+++ client/Player.js
@@ -20,3 +20,4 @@ const Player = (props) => {
 }
 
 export default Player
+
